Restaurant service Endpoints, Order Microservice creation, Kafka configuration

+++ restaurant_service/src/main/java/org/example/restaurant_service/Service/Impl/MenuItemServiceImpl.java
@@ -3,17 +3,22 @@ package org.example.restaurant_service.Service.Impl;
 import lombok.RequiredArgsConstructor;
 import org.example.restaurant_service.Domain.model.Category;
 import org.example.restaurant_service.Domain.model.MenuItem;
+import org.example.restaurant_service.Exception.ForbiddenAction;
+import org.example.restaurant_service.Exception.ResourceNotFound;
 import org.example.restaurant_service.Repository.MenuItemRepository;
 import org.example.restaurant_service.Service.MenuItemService;
+import org.example.restaurant_service.Service.Producer.MenuItemEventProducer;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @Service
 public class MenuItemServiceImpl implements MenuItemService {
     private final MenuItemRepository menuItemRepository;
+    private final MenuItemEventProducer menuItemEventProducer;
 
     @Override
     public MenuItem create(Long restaurantId, String name, BigDecimal price, Category category) {
@@ -22,13 +27,23 @@ public class MenuItemServiceImpl implements MenuItemService {
         menuItem.setName(name);
         menuItem.setPrice(price);
         menuItem.setCategory(category);
-        return menuItemRepository.save(menuItem);
+
+        menuItem =  menuItemRepository.save(menuItem);
+
+        menuItemEventProducer.createMenuItem(menuItem);
+        return menuItem;
     }
 
     @Override
-    public MenuItem update(Long id, String name, BigDecimal price) {
-        MenuItem menuItem = new MenuItem();
-        menuItem.setId(id);
+    public MenuItem update(Long id, String name, BigDecimal price, Category category, Long userId) {
+        MenuItem menuItem = this.findById(id);
+        if  (menuItem == null) {
+            throw new ResourceNotFound("Item not found");
+        }
+        if (!menuItem.getRestaurantId().equals(userId)) {
+            throw new ForbiddenAction("Not allowed to update item");
+        }
+        menuItem.setCategory(category);
         menuItem.setName(name);
         menuItem.setPrice(price);
         return menuItemRepository.save(menuItem);
@@ -40,8 +55,14 @@ public class MenuItemServiceImpl implements MenuItemService {
     }
 
     @Override
-    public MenuItem deleteById(Long id) {
+    public MenuItem deleteById(Long id,  Long userId) {
         MenuItem removedItem = this.findById(id);
+        if (removedItem == null) {
+            throw new ResourceNotFound("Item not found");
+        }
+        if(!removedItem.getRestaurantId().equals(userId)) {
+            throw new ForbiddenAction("Not allowed to delete item");
+        }
         menuItemRepository.delete(removedItem);
 
         return removedItem;
@@ -55,5 +76,15 @@ public class MenuItemServiceImpl implements MenuItemService {
     @Override
     public List<MenuItem> findAll() {
         return menuItemRepository.findAll();
+    }
+
+    @Override
+    public List<MenuItem> findByRestaurant(Long restaurantId) {
+        return menuItemRepository.findByRestaurantId(restaurantId);
+    }
+
+    @Override
+    public List<MenuItem> findMyRestaurantItems(Long restaurantId) {
+        return menuItemRepository.findByRestaurantId(restaurantId);
     }
 }
